python/requirements.txt
Pillow>=8.4
